Rename VpsIp to VpsIpSettings & VpsDdosList to VpsIpDdos
