+++ src/init/router.js
@@ -7,8 +7,8 @@ import Login from '../views/Login.vue'
 import Register from '../views/Register.vue'
 import Services from '../views/Services.vue'
 import Vps from '../views/Vps.vue'
-import VpsIp from '../views/VpsIp.vue'
-import VpsDdosList from '../views/VpsDdosList.vue'
+import VpsIpSettings from '../views/VpsIpSettings.vue'
+import VpsIpDdos from '../views/VpsIpDdos.vue'
 import Tickets from '../views/Tickets.vue'
 import Ticket from '../views/Ticket.vue'
 import Payments from '../views/Payments.vue'
@@ -32,8 +32,8 @@ export default new Router({
       //auth needed
       { path: '/:lg/service', component: Services},
       { path: '/:lg/service/vps/:id', component: Vps},
-      { path: '/:lg/service/vps/:id/ip/:ip', component: VpsIp},
-      { path: '/:lg/service/vps/:id/ip/:ip/ddos', component: VpsDdosList},
+      { path: '/:lg/service/vps/:id/ip/:ip/settings', component: VpsIpSettings},
+      { path: '/:lg/service/vps/:id/ip/:ip/ddos', component: VpsIpDdos},
       { path: '/:lg/ticket', component: Tickets},
       { path: '/:lg/ticket/:id', component: Ticket},
       { path: '/:lg/payment', component: Payments},
